move utility functions to lib directory
